refactor(rust): move answers-related code to agama-lib
